+++ llm_guard/input_scanners/secrets_plugins/adobe.py
@@ -1,0 +1,21 @@
+"""
+This plugin searches for Adobe keys
+"""
+import re
+
+from detect_secrets.plugins.base import RegexBasedDetector
+
+
+class AdobeSecretDetector(RegexBasedDetector):
+    """Scans for Adobe client keys."""
+
+    secret_type = "Adobe Client Keys"
+
+    denylist = [
+        # Adobe Client ID (OAuth Web)
+        re.compile(
+            r"""(?i)(?:adobe)(?:[0-9a-z\-_\t .]{0,20})(?:[\s|']|[\s|"]){0,3}(?:=|>|:{1,3}=|\|\|:|<=|=>|:|\?=)(?:'|\"|\s|=|\x60){0,5}([a-f0-9]{32})(?:['|\"|\n|\r|\s|\x60|;]|$)"""
+        ),
+        # Adobe Client Secret
+        re.compile(r"(?i)\b((p8e-)(?i)[a-z0-9]{32})(?:['|\"|\n|\r|\s|\x60|;]|$)"),
+    ]