+++ llm_guard/input_scanners/secrets_plugins/algolia_api_key.py
@@ -1,0 +1,16 @@
+"""
+This plugin searches for Algolia API keys
+"""
+import re
+
+from detect_secrets.plugins.base import RegexBasedDetector
+
+
+class AlgoliaApiKeyDetector(RegexBasedDetector):
+    """Scans for Algolia API keys."""
+
+    secret_type = "Algolia API Key"
+
+    denylist = [
+        re.compile(r"""(?i)\b((LTAI)(?i)[a-z0-9]{20})(?:['|\"|\n|\r|\s|\x60|;]|$)"""),
+    ]