+++ llm_guard/input_scanners/secrets_plugins/github_token.py
@@ -12,6 +12,10 @@ class GitHubTokenCustomDetector(RegexBasedDetector):
     secret_type = "GitHub Token"
 
     denylist = [
+        # GitHub App/Personal Access/OAuth Access/Refresh Token
         # ref. https://github.blog/2021-04-05-behind-githubs-new-authentication-token-formats/
         re.compile(r"(?:ghp|gho|ghu|ghs|ghr)_[A-Za-z0-9_]{36}"),
+        # GitHub Fine-Grained Personal Access Token
+        re.compile(r"github_pat_[0-9a-zA-Z_]{82}"),
+        re.compile(r"gho_[0-9a-zA-Z]{36}"),
     ]